refactor: 修改成异步
