--- OcrOnnxForm/packages.config
@@ -1,9 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Emgu.CV" version="4.4.0.4099" targetFramework="net461" />
-  <package id="Emgu.CV.Bitmap" version="4.4.0.4099" targetFramework="net461" />
-  <package id="System.Drawing.Common" version="4.7.0" targetFramework="net461" />
-  <package id="System.Drawing.Primitives" version="4.3.0" targetFramework="net461" />
-  <package id="System.Runtime" version="4.3.1" targetFramework="net461" />
-  <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.3.0" targetFramework="net461" />
-</packages>
