--- RapidOcrOnnxCs.sln
@@ -1,41 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.32929.386
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OcrOnnxForm", "OcrOnnxForm\OcrOnnxForm.csproj", "{43A96152-8891-4505-8389-32276F3A9422}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OcrLib", "OcrLib\OcrLib.csproj", "{AB10D356-2D95-4460-96CA-9A7129AD0A76}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{43A96152-8891-4505-8389-32276F3A9422}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{43A96152-8891-4505-8389-32276F3A9422}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{43A96152-8891-4505-8389-32276F3A9422}.Debug|x64.ActiveCfg = Debug|x64
-		{43A96152-8891-4505-8389-32276F3A9422}.Debug|x64.Build.0 = Debug|x64
-		{43A96152-8891-4505-8389-32276F3A9422}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{43A96152-8891-4505-8389-32276F3A9422}.Release|Any CPU.Build.0 = Release|Any CPU
-		{43A96152-8891-4505-8389-32276F3A9422}.Release|x64.ActiveCfg = Release|x64
-		{43A96152-8891-4505-8389-32276F3A9422}.Release|x64.Build.0 = Release|x64
-		{AB10D356-2D95-4460-96CA-9A7129AD0A76}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AB10D356-2D95-4460-96CA-9A7129AD0A76}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AB10D356-2D95-4460-96CA-9A7129AD0A76}.Debug|x64.ActiveCfg = Debug|x64
-		{AB10D356-2D95-4460-96CA-9A7129AD0A76}.Debug|x64.Build.0 = Debug|x64
-		{AB10D356-2D95-4460-96CA-9A7129AD0A76}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AB10D356-2D95-4460-96CA-9A7129AD0A76}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AB10D356-2D95-4460-96CA-9A7129AD0A76}.Release|x64.ActiveCfg = Release|x64
-		{AB10D356-2D95-4460-96CA-9A7129AD0A76}.Release|x64.Build.0 = Release|x64
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {6B47A7DB-6FE4-4943-9806-6D54D2BEC392}
-	EndGlobalSection
-EndGlobal
